+++ SupabaseProxy.Application/Interfaces/IPermissionRepository.cs
@@ -1,0 +1,8 @@
+using SupabaseProxy.Domain.Entities;
+
+namespace SupabaseProxy.Application.Interfaces;
+
+public interface IPermissionRepository
+{
+    Task<IEnumerable<Permission>> GetAllAsync();
+}