+++ SupabaseProxy.Application/Interfaces/IRoleRepository.cs
@@ -5,5 +5,6 @@ namespace SupabaseProxy.Application.Interfaces;
 public interface IRoleRepository
 {
     Task<IEnumerable<Role>> GetAllAsync();
+    Task<Role?> GetByIdAsync(Guid id);
     Task<Role?> GetByNameAsync(string name);
 }